Move prototype kit task into `compile-configs.mjs`
