+++ gulpfile.mjs
@@ -6,11 +6,11 @@ import taskListing from 'gulp-task-listing'
 import { paths, pkg } from './config/index.js'
 import { updateAssetsVersion } from './tasks/asset-version.mjs'
 import { clean } from './tasks/clean.mjs'
+import { updatePrototypeKitConfig } from './tasks/compile-configs.mjs'
 import { compileJavaScripts } from './tasks/compile-javascripts.mjs'
 import { compileStylesheets } from './tasks/compile-stylesheets.mjs'
 import { copyAssets, copyFiles } from './tasks/gulp/copy-to-destination.mjs'
 import { watch } from './tasks/gulp/watch.mjs'
-import { updatePrototypeKitConfig } from './tasks/prototype-kit-config.mjs'
 import { npmScriptTask } from './tasks/run.mjs'
 
 /**
