+++ tasks/compile-configs.mjs
